Added Fragment Support
Create new tests to verify that the requested fields are included in the $project

This was done by modifying the `getSelections()` function to use a reduce function over map.
This allowed to check if `sel.kind === 'FragmentSpread'`; if so, perform another recursive call to `getSelections`
A second argument was added to `getSelections`, which is the fragments of the incoming query.

The tests go 3 levels deep of fragments.